test: add windows server daemon smoke
