+++ cmd/deck/server_daemon_windows.go
@@ -5,9 +5,14 @@ package main
 import (
 	"os"
 	"os/exec"
+	"syscall"
 )
 
-func configureDetachedServerProcess(_ *exec.Cmd) {}
+const detachedProcess = 0x00000008
+
+func configureDetachedServerProcess(command *exec.Cmd) {
+	command.SysProcAttr = &syscall.SysProcAttr{CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP | detachedProcess}
+}
 
 func terminateDetachedServerProcess(pid int) error {
 	process, err := os.FindProcess(pid)
